Ho provato su Windows dal lavoro, sistemato alcune cose e tutto pare funzionare!!

+++ engine/src/external/tinyxml/tinyxml2.h
@@ -2315,7 +2315,7 @@ public:
     	includes the terminating null.)
     */
     int CStrSize() const {
-        return _buffer.Size();
+        return (int)_buffer.Size();
     }
     /**
     	If in print to memory mode, reset the buffer to the
